Bug #210116 - [WEB-UI] App Drawer is not showing properly (#1004)
* Bug #210116

* fix:prerak drawer

---------
