+++ lib/common-lib/src/components/layout/AppBar.js
@@ -12,29 +12,22 @@ import {
   InputRightAddon,
   Button,
   Modal,
-  VStack,
-  SmallCloseIcon,
-  Progress,
-  ChevronRightIcon,
-  Image,
-  Divider,
-  Alert
+  Image
+  // Divider
 } from 'native-base'
-import Drawer from 'react-modern-drawer'
 import 'react-modern-drawer/dist/index.css'
 import { useNavigate } from 'react-router-dom'
 import IconByName from '../IconByName'
 import { BodyLarge } from './HeaderTags'
-import { arrList, logout } from '../helper'
+import { logout } from '../helper'
 import { useTranslation } from 'react-i18next'
 import ImageView from '../ImageView'
 import * as FrontEndTypo from '../frontend_component'
-import { ChipStatus } from '../frontend_component/Chip'
 import { changeLanguage } from 'i18next'
-import { objProps } from '../helper'
 import Banner from '../Banner'
 
 export default function AppBar({
+  setIsOpen,
   isEnableSearchBtn,
   setSearch,
   setSearchState,
@@ -52,12 +45,12 @@ export default function AppBar({
   _backBtn,
   _menuBtn,
   _hstack,
-  facilitator,
   ...props
 }) {
   //fetch URL data and store fix for 2 times render useEffect call
   const [countLoad, setCountLoad] = useState(0)
   const [selectedCohortData, setSelectedCohortData] = useState({})
+  const [language, setLanguage] = useState()
 
   useEffect(() => {
     async function fetchData() {
@@ -80,24 +73,6 @@ export default function AppBar({
 
   const { t } = useTranslation()
   const [searchInput, setSearchInput] = React.useState(false)
-  const [isOpen, setIsOpen] = React.useState(false)
-  const [progress, setProgress] = useState(0)
-  const [status, setStatus] = React.useState('')
-  const [language, setLanguage] = React.useState()
-  const [prerakID, setPrerakId] = React.useState()
-
-  useEffect(() => {
-    try {
-      const prerakId = localStorage.getItem('id')
-      setPrerakId(prerakId)
-    } catch (error) {
-      console.error('Failed to fetch prerakID from localStorage:', error)
-    }
-  }, [])
-  const toggleDrawer = () => {
-    setIsOpen((prevState) => !prevState)
-  }
-
   const headerName = name
     ? name
     : localStorage.getItem('fullName')
@@ -147,35 +122,6 @@ export default function AppBar({
     changeLanguage(dLang ? dLang : 'en')
   }, [])
 
-  const res = objProps(facilitator)
-
-  React.useEffect(() => {
-    setProgress(
-      arrList(
-        {
-          ...res,
-          qua_name: facilitator?.qualifications?.qualification_master?.name
-        },
-        [
-          'device_ownership',
-          'mobile',
-          'device_type',
-          'gender',
-          'marital_status',
-          'social_category',
-          'name',
-          'contact_number',
-          'availability',
-          'aadhar_no',
-          'aadhaar_verification_mode',
-          'aadhar_verified',
-          'qualification_ids',
-          'qua_name'
-        ]
-      )
-    )
-    setStatus(facilitator?.status)
-  }, [])
   const handleImagePress = () => {
     navigate('/')
   }
@@ -447,225 +393,12 @@ export default function AppBar({
             )}
 
             {isShow('menuBtn', false) && (
-              <Box>
-                <HStack space={2} alignItems={'center'}>
-                  <IconByName
-                    name='MenuLineIcon'
-                    onPress={() => {
-                      setIsOpen(true)
-                    }}
-                  />
-                  <FrontEndTypo.H3>
-                    {selectedCohortData?.academic_year_name}
-                  </FrontEndTypo.H3>
-                </HStack>
-                <Drawer
-                  open={isOpen}
-                  onClose={toggleDrawer}
-                  direction='left'
-                  size='285px'
-                  style={{ zIndex: 99999 }}
-                >
-                  <Stack>
-                    <Box>
-                      <VStack paddingTop='10px' paddingRight='5'>
-                        <SmallCloseIcon
-                          size='sm'
-                          paddingLeft='250px'
-                          onClick={toggleDrawer}
-                          color='textMaroonColor.400'
-                        />
-
-                        <VStack alignItems='Center' display='flex'>
-                          {profile_url ? (
-                            <ImageView
-                              source={{
-                                uri: profile_url
-                              }}
-                              alt='Image not found'
-                              width={'132px'}
-                              height={'132px'}
-                            />
-                          ) : (
-                            <IconByName
-                              name='AccountCircleLineIcon'
-                              color='textGreyColor.600'
-                              _icon={{ size: '132' }}
-                            />
-                          )}
-                          <FrontEndTypo.H1
-                            pt='15px'
-                            color='textMaroonColor.400'
-                          >
-                            {headerName}
-                          </FrontEndTypo.H1>
-                          <ChipStatus
-                            status={
-                              facilitator ? facilitator?.status : 'under_review'
-                            }
-                          />
-                          <FrontEndTypo.H3 pt='15px'>
-                            {t('ID')}: {prerakID}
-                          </FrontEndTypo.H3>
-                        </VStack>
-                      </VStack>
-                    </Box>
-
-                    <Box pt='20px' gap='15px' pl='16px'>
-                      <FrontEndTypo.H2 bold='16px'>
-                        {progress !== 100
-                          ? t('COMPLETE_YOUR_PROFILE')
-                          : t('PROFILE_COMPLETED')}
-                      </FrontEndTypo.H2>
-                      <Box marginRight='10'>
-                        <Progress
-                          value={progress}
-                          size='xs'
-                          colorScheme='red'
-                        />
-                      </Box>
-                      <FrontEndTypo.H3>
-                        {(progress !== 100 && status === 'screened') ||
-                        (progress !== 100 && status === 'applied') ||
-                        (progress !== 100 &&
-                          status !== 'shortlisted_for_orientation')
-                          ? t('COMPLETE_THIS_BEFORE_ORIENTATION')
-                          : ''}
-                      </FrontEndTypo.H3>
-                      {/* </Pressable> */}
-
-                      <VStack space={2}>
-                        <Pressable
-                          onPress={(e) => {
-                            navigate('/profile')
-                          }}
-                          flex='1'
-                        >
-                          <HStack
-                            justifyContent='space-between'
-                            alignItems='center'
-                          >
-                            <HStack alignItems='Center' space='4'>
-                              <IconByName name='UserLineIcon' size='20px' />
-                              <FrontEndTypo.H2 ml='19px'>
-                                {t('PROFILE')}
-                              </FrontEndTypo.H2>
-                            </HStack>
-                            <ChevronRightIcon size='sm' color='textBlack.500' />
-                          </HStack>
-                        </Pressable>
-                        <Divider
-                          orientation='horizontal'
-                          background='dividerColor'
-                          width='268px'
-                          thickness='2'
-                        />
-                        <Pressable
-                          onPress={(e) => {
-                            navigate('/certificate')
-                          }}
-                          flex='1'
-                        >
-                          <HStack
-                            justifyContent='space-between'
-                            alignItems='center'
-                          >
-                            <HStack alignItems='Center' space='sm'>
-                              <IconByName name='ArticleLineIcon' size='20px' />
-                              <FrontEndTypo.H2 ml='19px'>
-                                {t('CERTIFICATE')}
-                              </FrontEndTypo.H2>
-                            </HStack>
-                            <ChevronRightIcon size='sm' color='textBlack.500' />
-                          </HStack>
-                        </Pressable>
-                        {/* <Divider
-                          orientation='horizontal'
-                          background='dividerColor'
-                          width='268px'
-                          thickness='1'
-                        />
-                        <HStack
-                          alignItems='Center'
-                          justifyContent='space-between'
-                        >
-                          <HStack alignItems='Center' space='sm'>
-                            <IconByName
-                              mt='5px'
-                              size='20px'
-                              name='Chat4LineIcon'
-                            />
-                            <FrontEndTypo.H2 ml='19px'>
-                              {t('FAQS')}
-                            </FrontEndTypo.H2>
-                          </HStack>
-                          <ChevronRightIcon size='sm' />
-                        </HStack>
-                        <Divider
-                          orientation='horizontal'
-                          background='dividerColor'
-                          width='268px'
-                          thickness='1'
-                        />
-                        <HStack
-                          alignItems='Center'
-                          justifyContent='space-between'
-                        >
-                          <HStack alignItems='Center' space='sm'>
-                            <IconByName name='PhoneLineIcon' size='20px' />
-                            <FrontEndTypo.H2 ml='19px'>
-                              {t('CALL_SUPPORT')}
-                            </FrontEndTypo.H2>
-                          </HStack>
-                          <ChevronRightIcon size='sm' />
-                        </HStack>
-                        <Divider
-                          orientation='horizontal'
-                          background='dividerColor'
-                          width='268px'
-                          thickness='1'
-                        />
-                        <HStack
-                          alignItems='Center'
-                          justifyContent='space-between'
-                        >
-                          <HStack alignItems='Center' space='sm'>
-                            <IconByName name='GroupLineIcon' size='20px' />
-                            <FrontEndTypo.H2 ml='19px'>
-                              {t('PRERAK_COMMUNITY')}
-                            </FrontEndTypo.H2>
-                          </HStack>
-                          <ChevronRightIcon size='sm' />
-                        </HStack>
-                        <Divider
-                          orientation='horizontal'
-                          background='dividerColor'
-                          width='268px'
-                          thickness='1'
-                        />
-                        <HStack
-                          alignItems='Center'
-                          justifyContent='space-between'
-                        >
-                          <HStack alignItems='Center' space='sm'>
-                            <IconByName name='Settings4LineIcon' size='20px' />
-                            <FrontEndTypo.H2 ml='19px'>
-                              {t('SETTINGS')}
-                            </FrontEndTypo.H2>
-                          </HStack>
-                          <ChevronRightIcon ml='52%' size='sm' />
-                        </HStack>
-                        <Divider
-                          orientation='horizontal'
-                          background='dividerColor'
-                          width='268px'
-                          thickness='1'
-                        /> */}
-                      </VStack>
-                    </Box>
-                  </Stack>
-                </Drawer>
-              </Box>
+              <IconByName
+                name='MenuLineIcon'
+                onPress={() => {
+                  setIsOpen(true)
+                }}
+              />
             )}
             {isShow('userInfo') && (
               <HStack space='1' alignItems='center'>
