+++ lib/common-lib/src/components/layout/Layout.js
@@ -6,6 +6,7 @@ import { getTokernUserInfo, useWindowSize } from '../helper'
 import HeightWidth from '../HeightWidth'
 import Loading from '../Loading'
 import { useLocation } from 'react-router-dom'
+import Drawer from '../layout/Drawer'
 
 export default function Layout({
   isDisabledAppBar,
@@ -19,12 +20,14 @@ export default function Layout({
   _appBar,
   _page,
   _center,
-  _footer
+  _footer,
+  facilitator
 }) {
   const [width, height] = useWindowSize()
   const [refFoot, setRefFoot] = React.useState()
   const [refAppBar, setRefAppBar] = React.useState()
   const location = useLocation()
+  const [isOpen, setIsOpen] = React.useState(false)
 
   React.useState(() => {
     const { resource_access } = getTokernUserInfo()
@@ -65,64 +68,66 @@ export default function Layout({
           }}
         />
       ) : (
-        <React.Fragment>
-          <HeightWidth
-            windowWidth='1080'
-            {..._page}
-            _scollView={{
-              shadow: 'appShadow',
-              ...(_page?._scollView ? _page?._scollView : {})
-            }}
-          >
-            <Stack
-              width={'100%'}
-              style={{
-                backgroundImage: imageUrl ? 'url(' + imageUrl + ')' : 'none',
-                backgroundColor: 'transparent',
-                backgroundRepeat: 'no-repeat',
-                backgroundSize: 'cover'
+        <HeightWidth>
+          <Drawer {...{ isOpen, setIsOpen, facilitator }} />
+          <React.Fragment>
+            <HeightWidth
+              windowWidth='1080'
+              {..._page}
+              _scollView={{
+                shadow: 'appShadow',
+                ...(_page?._scollView || {})
               }}
-              space={5}
             >
-              {!isDisabledAppBar ? (
+              <Stack
+                width={'100%'}
+                style={{
+                  backgroundImage: imageUrl ? 'url(' + imageUrl + ')' : 'none',
+                  backgroundColor: 'transparent',
+                  backgroundRepeat: 'no-repeat',
+                  backgroundSize: 'cover'
+                }}
+                space={5}
+              >
+                {!isDisabledAppBar && (
+                  <Box
+                    ref={(e) => {
+                      setRefAppBar(e)
+                      if (getRefAppBar) getRefAppBar(e)
+                    }}
+                  >
+                    <AppBar
+                      color={imageUrl ? 'white' : ''}
+                      languages={[
+                        { title: 'En', code: 'en' },
+                        { title: 'हिं', code: 'hi' }
+                      ]}
+                      {..._appBar}
+                      setIsOpen={setIsOpen}
+                    />
+                  </Box>
+                )}
+              </Stack>
+
+              {children}
+
+              {_footer?.menues && (
                 <Box
-                  ref={(e) => {
-                    setRefAppBar(e)
-                    if (getRefAppBar) getRefAppBar(e)
-                  }}
-                >
-                  <AppBar
-                    color={imageUrl ? 'white' : ''}
-                    languages={[
-                      { title: 'En', code: 'en' },
-                      { title: 'हिं', code: 'hi' }
-                    ]}
-                    {..._appBar}
-                  />
-                </Box>
-              ) : (
-                <React.Fragment />
+                  minH={
+                    refFoot?.clientHeight ? `${refFoot?.clientHeight}px` : 0
+                  }
+                />
               )}
-            </Stack>
-
-            {children}
-
-            {_footer?.menues ? (
-              <Box
-                minH={refFoot?.clientHeight ? `${refFoot?.clientHeight}px` : 0}
-              />
-            ) : (
-              <React.Fragment />
-            )}
-          </HeightWidth>
-          <Footer
-            setRef={(e) => {
-              setRefFoot(e)
-              if (getRefFoot) getRefFoot(e)
-            }}
-            {..._footer}
-          />
-        </React.Fragment>
+            </HeightWidth>
+            <Footer
+              setRef={(e) => {
+                setRefFoot(e)
+                if (getRefFoot) getRefFoot(e)
+              }}
+              {..._footer}
+            />
+          </React.Fragment>
+        </HeightWidth>
       )}
     </Center>
   )
